requirements.txt
django==4.1
flake8==4.0
python-dotenv==0.20.0
django-debug-toolbar~=3.5.0
MyApplication~=0.1.0
psycopg2~=2.9.3
Faker~=13.15.1